+++ src/Exceptions/OnOfficeQueryException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Innobrain\OnOfficeAdapter\Exceptions;
+
+use Exception;
+
+class OnOfficeQueryException extends Exception {}